Updates
Included/updated some data validation including live validation of inputs as user blurs out of each field. Fields are numeric data so check is done for that onBlur, check to see that fields such as GPA that are supposed to be floating point as as such and likewise with other fields that are integer. Fields in error turn red. Trimming errant blank rows on save assuming they were created in anticipation of adding another row of data that was never entered to keep the output clean. If an errorState is flagged (some invalid data encountered in input), save to output is canceled and user is alerted to make corrections.